precompute pivot inverse
This reduces the number of divisions, which are one of the most
expensive instructions (4 cycle throughput on Intel). While this does
not have a large effect at this (currently there are more than 16
instructions between two divs), it will enable better optimizations in
the future.

Note that with this change, the division unit drops from ~%36 activity
to ~%20 activity on colloc2 radial modes.