use column pivotting for the inner boundary to protect from numerical issues
If row pivotting is used, the equations from the inner boundary will get
mixed with the differential equations. In order to recover the inner
boundary condition after the full gaussian elimination, you would need
to have the full matrix. This is numerically bad, and will lead to the
boundary condition not being satisfied, as the matrices near the center
are ill-conditioned due to the 1/x term.

This is not an issue at the upper boundary, as the information is still
locally contained in just the last set of differential equations and the
boundary condition.

While the change in eigenfunctions is small (only a few points near the
center for a large (19000 point) model, it is relevant for mode
identification of at least the radial modes, introducing a spurious
crossing near the center (but not exactly at).